updated to use the new zvtm coordinates (doubles instead of longs)
git-svn-id: file:///home/hoijui/src/IsaViz/tmp2/zvtm.svn/code@3551 74e165f3-c697-40a6-91af-552c8936d7fe

+++ zvtm-cluster/trunk/src/main/java/fr/inria/zvtm/cluster/ClusteredImage.java
@@ -38,7 +38,7 @@ public class ClusteredImage extends VImage {
         DEFAULT_IMAGE = new BufferedImage(10,10,BufferedImage.TYPE_INT_ARGB);
     }
 
-    public ClusteredImage(long x, long y, int z, URL location, double scale){
+    public ClusteredImage(double x, double y, int z, URL location, double scale){
         super(x,y,z,DEFAULT_IMAGE,scale);
 
         this.location = location;
