+++ zvtm-cluster/trunk/src/main/java/fr/inria/zvtm/cluster/SlaveApp.java
@@ -145,12 +145,12 @@ public class SlaveApp {
 			return;
 		}
 
-		float focal = slaveCamera.getFocal();
-		float altCoef = (focal + masterLoc.alt) / focal;
+		double focal = slaveCamera.getFocal();
+		double altCoef = (focal + masterLoc.alt) / focal;
 		//block (screen) width in virtualspace coords
-		long virtBlockWidth = (long)(clusteredView.getClusterGeometry().getBlockWidth() * altCoef);
+		double virtBlockWidth = clusteredView.getClusterGeometry().getBlockWidth() * altCoef;
 		//block (screen) height in virtualspace coords
-		long virtBlockHeight = (long)(clusteredView.getClusterGeometry().getBlockHeight() * altCoef);	
+		double virtBlockHeight = clusteredView.getClusterGeometry().getBlockHeight() * altCoef;	
 
 		//row and col take origin block into account
 		int viewRows = clusteredView.getViewRows();
@@ -158,11 +158,11 @@ public class SlaveApp {
 		int row = clusteredView.rowNum(options.blockNumber) - clusteredView.rowNum(clusteredView.getOrigin()); 
 		int col = clusteredView.colNum(options.blockNumber) - clusteredView.colNum(clusteredView.getOrigin()); 
 
-		long xOffset = -((viewCols-1)*virtBlockWidth)/2; 
-		long yOffset = ((viewRows-1)*virtBlockHeight)/2;
+		double xOffset = -((viewCols-1)*virtBlockWidth)/2; 
+		double yOffset = ((viewRows-1)*virtBlockHeight)/2;
 
-		long newX = xOffset + masterLoc.vx + col*virtBlockWidth;
-		long newY = -yOffset + masterLoc.vy - row*virtBlockHeight;
+		double newX = xOffset + masterLoc.vx + col*virtBlockWidth;
+		double newY = -yOffset + masterLoc.vy - row*virtBlockHeight;
 
 		slaveCamera.setLocation(new Location(newX, newY, masterLoc.alt));
 	}
